refactor(benchmarks): Delete the wire mirrors, share one schema
The three-way fork is closed. `EnvFamily`, `FamilyPayload`,
`EpisodeRecord`, `FrameRecord`, `MetricSample`, the version constants,
`RecordChunk`, and `decode_episode_record` now live once, in
`rlevo-scene`, and the recorder and the WASM report client both use them.

Net 1,139 lines removed. The six `*Payload` structs in `schema.rs` -- each
documented as a "bincode-stable mirror" of a snapshot type, each with a
hand-written `From` -- are gone, and `FamilyPayload` carries the snapshots
directly, so `FamilyPayload::Grid(GridPayload::from(e.grid_snapshot()))`
is now `FamilyPayload::Grid(e.grid_snapshot())`. The 495-line drift guard
is deleted: it guarded two definitions, and there is one.

`RecordChunk` was forked too, which the ADR did not record -- a private
copy in the writer and a public one in the client, so the framing drifted
on the same terms as the payloads. Shared now.

Three things stay in `rlevo-benchmarks` because they are not wire data:
`RecordedEnvFamily` (a producer-side opt-in), `default_frame_stride`
(recording policy), and `RunId::new_now`, which becomes the free function
`new_run_id` -- the type is wire data but minting one needs a clock and an
RNG, and the leaf has neither by construction.

`wire.rs` drops 839 lines to 117: re-exports, plus `RunManifest` and
`ObjectiveSense`, which stay mirrored because the manifest crosses as JSON
and a field-keyed format degrades where a positional one corrupts. Its
docs now say so, and say not to finish the job.

CRITICAL: Unifying proved the mirror had never been a faithful copy, in
two ways the drift guard structurally could not catch -- it round-tripped
values, so it saw neither. `Modifier`'s field is private with a `bits()`
accessor; the mirror declared it `pub`, and client code read `.0` and
built `Modifier(1)` in tests. And the real enums are `#[non_exhaustive]`
while the mirrors were not, so eight matches across the client and the
report compiled as exhaustive and now need explicit arms. Every one of
those arms is a behaviour decision that did not previously exist: an
unknown variant now renders a documented fallback instead of failing to
compile. Reviewers should check those arms, not just that it builds.

Verified: workspace all-targets, fast tier 12/12, clippy clean but for one
pre-existing xtask nit, rustdoc clean at default *and* --all-features, and
both `rlevo-scene` and the report client still build for
wasm32-unknown-unknown.
