+++ crates/rlevo-benchmarks-report-client/src/wire.rs
@@ -1,649 +1,46 @@
-//! Bincode-compatible mirror of `rlevo_benchmarks::record::schema`.
+//! The run manifest, and the one thing the client still mirrors on purpose.
 //!
-//! The native record schema is gated behind `feature = "record"` in
-//! `rlevo-benchmarks`, which itself doesn't compile to `wasm32` because
-//! `rand`/`getrandom` need the `js` feature there. Rather than restructure
-//! the workspace, the client redeclares the wire types here. The
-//! cross-crate compatibility test in `rlevo-benchmarks` round-trips a
-//! populated `EpisodeRecord` between the two representations on every
-//! `cargo test` to guarantee they stay byte-identical.
+//! Everything that reaches the client as **bincode** — [`EpisodeRecord`],
+//! [`FrameRecord`], [`FamilyPayload`], the payload shapes, the styled-text
+//! types, [`FORMAT_VERSION`], and [`decode_episode_record`] — now comes from
+//! [`rlevo_scene`], defined once and shared with the recorder. This module
+//! re-exports it so `crate::wire::…` paths inside the client keep working.
 //!
-//! **Sync contract:** these struct/enum definitions MUST stay in
-//! lockstep with `rlevo-benchmarks/src/record/schema.rs`. Field order,
-//! field types, and `#[non_exhaustive]` markers all participate in the
-//! bincode wire format. Bumping `FORMAT_VERSION` is the canonical
-//! escape hatch when the contract changes.
+//! # Why `RunManifest` is still a copy
+//!
+//! **This is deliberate, not leftover.** Do not "finish the job" by deleting it.
+//!
+//! The manifest does not travel as bincode. `rlevo-benchmarks` emits it with
+//! `serde_json::to_string` into a `<script type="application/json">` block, and
+//! that difference decides everything:
+//!
+//! | Crosses as | Keyed by | Drift behaviour |
+//! |---|---|---|
+//! | bincode | position / variant tag | **silent corruption** — hence one definition |
+//! | JSON | field name | graceful: unknown ignored, missing become `None` |
+//!
+//! *Mirror what is self-describing; share what is positional.*
+//!
+//! Sharing it instead would mean moving [`ObjectiveSense`] — which the manifest
+//! carries — out of `rlevo-core`, where an optimisation concept belongs, and
+//! into a record crate. That is a worse trade than a 70-line struct whose worst
+//! failure mode is a field silently reading `None`. See ADR 0081.
+
+pub use rlevo_scene::codec::{DecodeError, RecordChunk, decode_episode_record};
+pub use rlevo_scene::payload::{
+    BodyKind, Box2dSnapshot, CardTable, Classic2DBody, Classic2DRole, Classic2DSnapshot,
+    GridAgentMarker, GridColor, GridDir, GridDoorState, GridSnapshot, GridTile,
+    Landscape2DSnapshot, Locomotion2DSnapshot, Point2, RigidBody2D, TabularCell, TabularGrid,
+    TabularLayout, TabularMarker, TabularMarkerKind, TabularSnapshot,
+};
+pub use rlevo_scene::schema::{
+    CheckpointFormat, CheckpointKind, CheckpointRef, EnvFamily, EpisodeKind, EpisodeRecord,
+    EpisodeRecordHeader, FORMAT_VERSION, FamilyPayload, FrameRecord, Hyperparameters,
+    MIN_SUPPORTED_VERSION, MetricSample, PopulationSample, RunId, TrialRef, bincode_config,
+};
+pub use rlevo_scene::styled::{Color, Modifier, SpanStyle, StyledFrame, StyledLine, StyledSpan};
 
 use serde::{Deserialize, Serialize};
-use std::collections::BTreeMap;
-
-// ---- Mirror of rlevo_scene::styled (kept thin; the cross-crate
-// compat test in rlevo-benchmarks guards every field). -----------------
-
-/// Wire mirror of `rlevo_scene::StyledFrame` — one complete rendered frame.
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct StyledFrame {
-    /// Ordered rows of the rendered frame, top to bottom.
-    pub lines: Vec<StyledLine>,
-}
-
-/// Wire mirror of `rlevo_scene::StyledLine` — one row of styled spans.
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct StyledLine {
-    /// Consecutive styled text segments that make up the row.
-    pub spans: Vec<StyledSpan>,
-}
-
-/// Wire mirror of `rlevo_scene::StyledSpan` — a run of identically-styled text.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct StyledSpan {
-    /// The rendered text content of this span.
-    pub text: String,
-    /// Visual style (foreground, background, modifier bits) to apply.
-    pub style: SpanStyle,
-}
-
-/// Wire mirror of `rlevo_scene::SpanStyle` — the visual attributes for one span.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct SpanStyle {
-    /// Optional foreground colour; `None` means inherit the terminal default.
-    pub fg: Option<Color>,
-    /// Optional background colour; `None` means inherit the terminal default.
-    pub bg: Option<Color>,
-    /// Bit-packed text modifier flags (bold, dim, italic, underlined, reversed).
-    pub modifier: Modifier,
-}
-
-/// Wire mirror of `rlevo_scene::Color` — the named terminal palette plus an indexed escape.
-///
-/// `Reset` and `Indexed(_)` map to no CSS class in the web renderer (see
-/// [`crate::styled::color_class`]).
-#[non_exhaustive]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum Color {
-    Reset,
-    Black,
-    Red,
-    Green,
-    Yellow,
-    Blue,
-    Magenta,
-    Cyan,
-    Gray,
-    DarkGray,
-    LightRed,
-    LightGreen,
-    LightYellow,
-    LightBlue,
-    LightMagenta,
-    LightCyan,
-    White,
-    /// 256-colour indexed terminal colour; rendered unstyled in the web report.
-    Indexed(u8),
-}
-
-/// Wire mirror of `rlevo_scene::Modifier` — bit-packed text decoration flags.
-///
-/// Bit layout: BOLD=1, DIM=2, ITALIC=4, UNDERLINED=8, REVERSED=16.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct Modifier(pub u8);
-
-// ---- /styled mirror ---------------------------------------------------
-
-// ---- Mirror of rlevo_scene::payload (rich per-family payloads). --
-
-/// A 2-D point in world coordinates, shared by all rich family payloads.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Serialize, Deserialize)]
-pub struct Point2 {
-    /// Horizontal component in world space.
-    pub x: f32,
-    /// Vertical component in world space.
-    pub y: f32,
-}
-
-impl Point2 {
-    /// Creates a new [`Point2`] with the given coordinates.
-    #[must_use]
-    pub const fn new(x: f32, y: f32) -> Self {
-        Self { x, y }
-    }
-}
-
-/// Semantic role of a [`RigidBody2D`] within a `Box2D` scene.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum BodyKind {
-    /// Main agent body (e.g. lander hull, walker torso).
-    Hull,
-    /// Circular wheel attached to a vehicle or walker.
-    Wheel,
-    /// Limb segment (bipedal walker leg).
-    Leg,
-    /// Wing surface (not yet used; reserved for future flying agents).
-    Wing,
-    /// Static terrain or floor polygon.
-    Ground,
-    /// Landing pad or target region.
-    Goal,
-    /// Any body that doesn't fit a named category.
-    Other,
-}
-
-/// Wire mirror of a single rigid body in a `Box2D` scene.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct RigidBody2D {
-    /// Local-frame polygon vertices, ordered for rendering.
-    pub vertices: Vec<Point2>,
-    /// World-space centre of the body at this frame.
-    pub position: Point2,
-    /// Counter-clockwise rotation from the body's rest orientation, in radians.
-    pub rotation_rad: f32,
-    /// Semantic role used to select the CSS class and legend entry.
-    pub kind: BodyKind,
-}
-
-/// Per-frame snapshot for a 2-D landscape-search environment.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Landscape2DPayload {
-    /// `(lo, hi)` search domain bounds along the x-axis.
-    pub bounds_x: (f32, f32),
-    /// `(lo, hi)` search domain bounds along the y-axis.
-    pub bounds_y: (f32, f32),
-    /// Position of the current candidate solution.
-    pub current: Point2,
-    /// Best candidate found so far, if available.
-    pub best: Option<Point2>,
-    /// Recent candidate positions, oldest first (used for the trail polyline).
-    pub trail: Vec<Point2>,
-    /// Human-readable landscape name (`"sphere"`, `"ackley"`, `"rastrigin"`, …).
-    pub label: String,
-}
-
-/// Per-frame snapshot for a `Box2D` physics environment.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Box2dPayload {
-    /// Axis-aligned bounding box of the visible world, as `(min, max)`.
-    pub world_bounds: (Point2, Point2),
-    /// All rigid bodies in the scene at this frame.
-    pub bodies: Vec<RigidBody2D>,
-    /// Active contact points between bodies (rendered as small open rings).
-    pub contacts: Vec<Point2>,
-}
-
-/// Per-frame snapshot for a 2-D locomotion environment (sagittal-plane view).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Locomotion2DPayload {
-    /// World-space positions of every skeleton joint.
-    pub joints: Vec<Point2>,
-    /// Pairs of joint indices defining bones; index into [`joints`](Self::joints).
-    pub bones: Vec<(u32, u32)>,
-    /// World-space y-coordinate of the ground plane, or `None` for an env
-    /// with no ground (top-down, zero-gravity envs such as `Swimmer` and
-    /// `Reacher`).
-    pub ground_y: Option<f32>,
-    /// Centre-of-mass position, if computed and available.
-    pub com: Option<Point2>,
-    /// Active contact points with the ground (rendered as open rings).
-    pub contacts: Vec<Point2>,
-}
-
-/// Cardinal facing of the grid agent (mirror of `rlevo-core` `GridDir`).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum GridDir {
-    East,
-    South,
-    West,
-    North,
-}
-
-/// The six Minigrid colours (mirror of `rlevo-core` `GridColor`).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum GridColor {
-    Red,
-    Green,
-    Blue,
-    Purple,
-    Yellow,
-    Grey,
-}
-
-/// Door state (mirror of `rlevo-core` `GridDoorState`).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum GridDoorState {
-    Open,
-    Closed,
-    Locked,
-}
-
-/// One grid cell's contents (mirror of `rlevo-core` `GridTile`).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum GridTile {
-    Empty,
-    Floor,
-    Wall,
-    Goal,
-    Lava,
-    Door(GridColor, GridDoorState),
-    Key(GridColor),
-    Ball(GridColor),
-    Box(GridColor),
-}
-
-/// Agent marker: cell position, facing, carried item (mirror of
-/// `rlevo-core` `GridAgentMarker`).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct GridAgentMarker {
-    /// Column index of the agent (0 = left edge).
-    pub x: u16,
-    /// Row index of the agent (0 = top edge).
-    pub y: u16,
-    /// Cardinal direction the agent is currently facing.
-    pub dir: GridDir,
-    /// Tile the agent is carrying, if any.
-    pub carrying: Option<GridTile>,
-}
-
-/// Structured tile grid for `grids` envs (mirror of
-/// `rlevo-benchmarks::record::GridPayload`). `tiles` is row-major,
-/// `len == width * height`; cell `(x, y)` is `tiles[y * width + x]`.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct GridPayload {
-    /// Number of columns in the grid.
-    pub width: u16,
-    /// Number of rows in the grid.
-    pub height: u16,
-    /// Row-major flat array of tile contents; length must equal `width * height`.
-    pub tiles: Vec<GridTile>,
-    /// Current position, facing, and carried item of the agent.
-    pub agent: GridAgentMarker,
-}
-
-/// Background class of a tabular grid cell (mirror of `rlevo-core`
-/// `TabularCell`).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum TabularCell {
-    Empty,
-    Frozen,
-    Start,
-    Goal,
-    Hazard,
-}
-
-/// Marker class overlaid on a tabular grid cell (mirror of `rlevo-core`
-/// `TabularMarkerKind`).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum TabularMarkerKind {
-    Agent,
-    Passenger,
-    Destination,
-    Location,
-}
-
-/// A point-of-interest overlaid on a tabular grid cell (mirror of
-/// `rlevo-core` `TabularMarker`).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct TabularMarker {
-    /// Column index of the cell (0 = left edge).
-    pub x: u16,
-    /// Row index of the cell (0 = top edge).
-    pub y: u16,
-    /// What kind of marker is overlaid on this cell.
-    pub kind: TabularMarkerKind,
-}
-
-/// Grid layout for grid-shaped toy-text envs (mirror of `rlevo-core`
-/// `TabularGrid`). `cells` is row-major, `len == width * height`.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct TabularGrid {
-    /// Number of columns in the grid.
-    pub width: u16,
-    /// Number of rows in the grid.
-    pub height: u16,
-    /// Row-major flat array of background cell types; length must equal `width * height`.
-    pub cells: Vec<TabularCell>,
-    /// Overlay markers (agent, passenger, destination, …) rendered on top of `cells`.
-    pub markers: Vec<TabularMarker>,
-}
-
-/// Card-table layout for Blackjack (mirror of `rlevo-core` `CardTable`).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct CardTable {
-    /// Cards held by the player (face values 1–10).
-    pub player_cards: Vec<u8>,
-    /// Current sum of the player's hand.
-    pub player_total: u8,
-    /// Whether the player holds a usable ace (counts as 11 without busting).
-    pub usable_ace: bool,
-    /// All dealer cards, including the hidden hole card (revealed at episode end).
-    pub dealer_cards: Vec<u8>,
-    /// The single dealer card visible to the player during play.
-    pub dealer_showing: u8,
-}
-
-/// Layout discriminant (mirror of `rlevo-core` `TabularLayout`).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum TabularLayout {
-    Grid(TabularGrid),
-    Cards(CardTable),
-}
-
-/// Structured toy-text layout (mirror of `rlevo-benchmarks::record::TabularPayload`).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct TabularPayload {
-    /// Discriminated layout: either a tile grid or a card table.
-    pub layout: TabularLayout,
-}
-
-/// Semantic role of a classic-control body (mirror of `rlevo-core`
-/// `Classic2DRole`).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum Classic2DRole {
-    Track,
-    Cart,
-    Pole,
-    Link,
-    Car,
-    Hinge,
-}
-
-/// One world-space body of a classic-control mechanism (mirror of
-/// `rlevo-core` `Classic2DBody`).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Classic2DBody {
-    /// World-space vertices of the polyline or polygon, in drawing order.
-    pub points: Vec<Point2>,
-    /// Semantic role; used to select the CSS class for the web renderer.
-    pub role: Classic2DRole,
-    /// When `true`, connect the last point back to the first (polygon); otherwise render as an open polyline.
-    pub closed: bool,
-}
-
-/// Structured 2-D line-art for classic-control envs (mirror of
-/// `rlevo-benchmarks::record::Classic2DPayload`).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Classic2DPayload {
-    /// All bodies (track, cart, pole, …) that make up the scene.
-    pub bodies: Vec<Classic2DBody>,
-    /// Axis-aligned bounding box of the visible world, as `(min, max)`.
-    pub bounds: (Point2, Point2),
-}
-
-// ---- /payload mirror --------------------------------------------------
-
-/// Current wire-format version this client crate writes/expects.
-///
-/// # Version history
-///
-/// | Version | Changes |
-/// |---------|---------|
-/// | 1 | Initial format: header, frames, metrics. |
-/// | 2 | Added [`PopulationSample`] / [`RecordChunk::Population`]. |
-/// | 3 | Added `diversity`, `best_index`, `best_genome_digest`, `parents_of_best`, `inner_rl_returns` to [`PopulationSample`]. |
-/// | 4 | Added [`TrialRef`] (`trial` field) to [`EpisodeRecordHeader`]. |
-/// | 5 | Added [`FamilyPayload::Grid`], [`FamilyPayload::TabularText`], [`FamilyPayload::Classic2D`] rich payloads. |
-/// | 6 | Added [`EpisodeKind`] to [`EpisodeRecordHeader`]; added run-provenance fields (`algorithm`, `rlevo_version`, `rustc_version`, `burn_version`, `platform`, `git_commit`, `git_dirty`, `device`, `num_seeds`, `success_threshold`) and `checkpoints` ([`CheckpointRef`]) to [`RunManifest`]. Defined by ADR 0014. |
-/// | 7 | Added [`ObjectiveSense`] and the `objective_sense` field to [`RunManifest`]; orients best/worst report transforms. |
-/// | 8 | [`Locomotion2DPayload::ground_y`] became `Option<f32>`. `Swimmer` and `Reacher` are top-down, zero-gravity envs with no ground plane; a bare `f32` forced them to name a floor that does not exist and the adapter drew it through the figure. |
-// Mirror of `rlevo-benchmarks::record::FORMAT_VERSION`.  Keep in sync;
-// the const assertions in rlevo-benchmarks/tests/wire_format_compat.rs
-// catch drift at compile time.
-pub const FORMAT_VERSION: u16 = 8;
-
-/// Oldest on-disk version this client accepts. Equal to
-/// [`FORMAT_VERSION`] — no backward compatibility before first release.
-// Mirror of `rlevo-benchmarks::record::MIN_SUPPORTED_VERSION`.
-pub const MIN_SUPPORTED_VERSION: u16 = 8;
-
-/// Returns the standard bincode configuration used for all record encode/decode operations.
-#[must_use]
-pub fn bincode_config() -> bincode::config::Configuration {
-    bincode::config::standard()
-}
-
-/// Opaque string identifier for a training run, unique within a workspace.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RunId(pub String);
-
-/// Discriminant for the environment family recorded in a run.
-///
-/// `#[non_exhaustive]` — new families may be added in future wire-format
-/// versions without breaking existing clients; match arms should include a
-/// wildcard that routes to [`crate::adapters::fallback`].
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum EnvFamily {
-    /// `CartPole`, `MountainCar`, Pendulum, Acrobot.
-    Classic,
-    /// Minigrid-style grid worlds (empty, four-rooms, door-key, …).
-    Grids,
-    /// `FrozenLake`, `CliffWalking`, Taxi, Blackjack.
-    ToyText,
-    /// `LunarLander`, `BipedalWalker`, `CarRacing` (`Box2D` physics).
-    Box2d,
-    /// Ant, `HalfCheetah`, Hopper, `Walker2D` (`MuJoCo` / locomotion).
-    Locomotion,
-    /// Sphere, Ackley, Rastrigin (black-box optimisation landscapes).
-    Landscapes,
-}
-
-/// Family-specific per-frame data appended to each [`FrameRecord`].
-///
-/// `Ascii` is the fallback when the environment has no rich payload.
-/// `#[non_exhaustive]` — new variants may arrive in future format versions.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum FamilyPayload {
-    /// No rich payload; the frame carries only ASCII / styled text.
-    Ascii,
-    /// 2-D landscape-search snapshot (Sphere, Ackley, Rastrigin, …).
-    Landscape2D(Landscape2DPayload),
-    /// `Box2D` rigid-body scene (`LunarLander`, `BipedalWalker`, …).
-    Box2dBodies(Box2dPayload),
-    /// Sagittal-plane locomotion skeleton (Ant, `HalfCheetah`, …).
-    Locomotion2D(Locomotion2DPayload),
-    /// Structured tile grid (empty, four-rooms, door-key, …). Added in
-    /// `FORMAT_VERSION = 5`.
-    Grid(GridPayload),
-    /// Structured toy-text layout (FrozenLake/CliffWalking/Taxi grid, or
-    /// Blackjack cards). Added in `FORMAT_VERSION = 5`.
-    TabularText(TabularPayload),
-    /// Structured 2-D line-art for classic physics envs. Added in
-    /// `FORMAT_VERSION = 5`.
-    Classic2D(Classic2DPayload),
-}
-
-/// Objective direction for a run. Wire mirror of
-/// `rlevo_core::objective::ObjectiveSense`. `None` on a [`RunManifest`] is
-/// interpreted as [`Maximize`](ObjectiveSense::Maximize) — the canonical
-/// engine sense. Added in `FORMAT_VERSION = 7`.
-// Mirror of `rlevo_core::objective::ObjectiveSense`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum ObjectiveSense {
-    /// Lower is better (cost, loss, error, distance-to-target).
-    Minimize,
-    /// Higher is better (reward, fitness, accuracy, score).
-    Maximize,
-}
-
-/// Trial provenance. Mirror of
-/// `rlevo_benchmarks::record::schema::TrialRef`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct TrialRef {
-    /// Zero-based environment index within the suite.
-    pub env_index: u32,
-    /// Zero-based seed-repetition index for that environment.
-    pub trial_index: u32,
-}
-
-/// Whether an episode was a training or evaluation rollout. Added in
-/// `FORMAT_VERSION = 6`.
-// Mirror of `rlevo-benchmarks::record::EpisodeKind`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum EpisodeKind {
-    /// Exploration rollout produced while the policy is being updated.
-    #[default]
-    Training,
-    /// Deterministic rollout produced to measure the current policy.
-    Evaluation,
-}
-
-/// Why a [`CheckpointRef`] was written. Added in `FORMAT_VERSION = 6`.
-// Mirror of `rlevo-benchmarks::record::CheckpointKind`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum CheckpointKind {
-    /// Saved on a periodic step interval during training.
-    Periodic,
-    /// Saved because it is the best-performing checkpoint so far.
-    Best,
-    /// Saved at the end of the run.
-    Final,
-}
-
-/// Burn recorder format of a saved learner checkpoint. Added in
-/// `FORMAT_VERSION = 6`.
-// Mirror of `rlevo-benchmarks::record::CheckpointFormat`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum CheckpointFormat {
-    /// `NamedMpkFileRecorder` (`.mpk`).
-    NamedMpk,
-    /// `NamedMpkGzFileRecorder` (`.mpk.gz`).
-    NamedMpkGz,
-    /// `BinFileRecorder` (`.bin`).
-    Bin,
-    /// `PrettyJsonFileRecorder` (`.json`).
-    Json,
-    /// Any other recorder.
-    Other,
-}
-
-/// Reference to one Burn-saved learner checkpoint; the record references
-/// the file and never embeds weights. Added in `FORMAT_VERSION = 6`.
-// Mirror of `rlevo-benchmarks::record::CheckpointRef`.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct CheckpointRef {
-    /// Global training step at which the checkpoint was written.
-    pub step: u32,
-    /// Why this checkpoint was written.
-    pub kind: CheckpointKind,
-    /// Burn recorder format used for the artifact.
-    pub format: CheckpointFormat,
-    /// Path to the artifact, relative to the run directory.
-    pub path: String,
-    /// Metric (e.g. eval return) at save time, if known.
-    pub metric: Option<f64>,
-    /// 128-bit content digest.
-    pub digest: Option<[u8; 16]>,
-}
-
-/// Fixed-size preamble written at the start of every `.rec` file.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct EpisodeRecordHeader {
-    /// Wire-format version; must equal [`FORMAT_VERSION`] to be decoded.
-    pub format_version: u16,
-    /// Identifier of the training run that produced this file.
-    pub run_id: RunId,
-    /// RNG seed used for the run (for reproducibility).
-    pub seed: u64,
-    /// Environment family; controls which family adapter renders the frames.
-    pub env_family: EnvFamily,
-    /// Unix timestamp (seconds) when the episode file was created.
-    pub created_at: i64,
-    /// Trial that produced this episode, or `None` for non-harness
-    /// producers. Added in `FORMAT_VERSION = 4`.
-    pub trial: Option<TrialRef>,
-    /// Whether this episode was a training or evaluation rollout. Added in
-    /// `FORMAT_VERSION = 6`.
-    pub kind: EpisodeKind,
-}
-
-/// One recorded simulation step, stored as a [`RecordChunk::Frame`].
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct FrameRecord {
-    /// Simulation step counter within the episode.
-    pub step: u32,
-    /// Raw action bytes serialised by the agent (interpretation is family-specific).
-    pub action: Vec<u8>,
-    /// Scalar reward received at this step.
-    pub reward: f32,
-    /// Plain-text ASCII render produced by the library tier, if available.
-    pub ascii: Option<String>,
-    /// Colour-annotated styled render produced by the library tier, if available.
-    pub styled: Option<StyledFrame>,
-    /// Family-specific rich payload (bodies, joints, landscape position, …).
-    pub family_payload: FamilyPayload,
-}
-
-/// A single named scalar measurement emitted during training, stored as [`RecordChunk::Metrics`].
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct MetricSample {
-    /// Training step at which the metric was recorded.
-    pub step: u32,
-    /// Metric key (e.g. `"policy_loss"`, `"entropy"`, `"best_fitness"`).
-    pub name: String,
-    /// Scalar value of the metric at `step`.
-    pub value: f64,
-}
-
-/// One per-generation snapshot of an EA population, stored as [`RecordChunk::Population`].
-///
-/// Mirror of `rlevo_benchmarks::record::PopulationSample`; present in v3+ records.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct PopulationSample {
-    /// EA generation index; used as the x-axis in population panels.
-    pub generation: u32,
-    /// Raw fitness score for every individual in the population, in population order.
-    pub fitnesses: Vec<f32>,
-    /// Optional scalar diversity metric (e.g. mean pairwise genome distance).
-    pub diversity: Option<f32>,
-    /// Index into `fitnesses` of the best individual this generation.
-    pub best_index: u32,
-    /// 128-bit content digest of the best genome (for reproducibility); `None` when not recorded.
-    pub best_genome_digest: Option<[u8; 16]>,
-    /// Content digests of the parents that produced the best genome.
-    pub parents_of_best: Vec<[u8; 16]>,
-    /// Per-individual RL episode returns when the outer EA wraps an inner RL loop.
-    pub inner_rl_returns: Option<Vec<f32>>,
-}
-
-/// Length-prefixed wire-format chunk written by the on-disk record writer.
-///
-/// **Variant ordering is wire-format-stable** — new variants append at
-/// the end so existing bincode tags keep decoding. `Population` is at tag 2.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum RecordChunk {
-    /// One simulation step (bincode tag 0).
-    Frame(FrameRecord),
-    /// Batch of metric samples emitted together (bincode tag 1).
-    Metrics(Vec<MetricSample>),
-    /// One EA population snapshot, present in v2+ records (bincode tag 2).
-    Population(PopulationSample),
-}
-
-/// In-memory aggregate of one fully decoded `.rec` file.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct EpisodeRecord {
-    /// Fixed-size preamble with run identity and format metadata.
-    pub header: EpisodeRecordHeader,
-    /// Ordered sequence of recorded simulation steps.
-    pub frames: Vec<FrameRecord>,
-    /// All metric samples emitted during the episode, in emission order.
-    pub metrics: Vec<MetricSample>,
-    /// Per-generation EA population snapshots; empty for RL-only runs.
-    #[serde(default)]
-    pub population_samples: Vec<PopulationSample>,
-}
-
-/// Key-value map of algorithm hyperparameters recorded at run start.
-pub type Hyperparameters = BTreeMap<String, String>;
 
 /// JSON manifest embedded in `index.html` describing the overall run.
 ///
@@ -711,129 +108,10 @@ pub struct RunManifest {
     pub objective_sense: Option<ObjectiveSense>,
 }
 
-/// Decode the raw bytes of a single `episode_*.rec` file produced by
-/// the on-disk record writer. Tolerates truncated tails by stopping
-/// cleanly at the last whole chunk.
-///
-/// # Errors
-///
-/// Returns [`DecodeError::Truncated`] if the preamble is shorter than
-/// 16 bytes, [`DecodeError::VersionMismatch`] if the format version in
-/// the preamble does not match [`FORMAT_VERSION`], or
-/// [`DecodeError::Bincode`] if a length-prefixed chunk fails to
-/// deserialise.
-pub fn decode_episode_record(bytes: &[u8]) -> Result<EpisodeRecord, DecodeError> {
-    if bytes.len() < 16 {
-        return Err(DecodeError::Truncated("preamble"));
-    }
-    let version = u16::from_le_bytes([bytes[0], bytes[1]]);
-    if version != FORMAT_VERSION {
-        return Err(DecodeError::VersionMismatch {
-            file: version,
-            client: FORMAT_VERSION,
-        });
-    }
-    let mut cursor = 16;
-    let header: EpisodeRecordHeader =
-        read_chunk(bytes, &mut cursor)?.ok_or(DecodeError::Truncated("header"))?;
-
-    let mut frames = Vec::new();
-    let mut metrics = Vec::new();
-    let mut population_samples = Vec::new();
-    while let Some(chunk) = read_chunk::<RecordChunk>(bytes, &mut cursor)? {
-        match chunk {
-            RecordChunk::Frame(fr) => frames.push(fr),
-            RecordChunk::Metrics(ms) => metrics.extend(ms),
-            RecordChunk::Population(ps) => population_samples.push(ps),
-        }
-    }
-    Ok(EpisodeRecord {
-        header,
-        frames,
-        metrics,
-        population_samples,
-    })
-}
-
-/// Reads one length-prefixed bincode chunk from `bytes` at `*cursor`.
-///
-/// Advances `*cursor` past the 4-byte length prefix and the payload.
-/// Returns `Ok(None)` on a truncated length prefix or partial payload so
-/// the caller can stop cleanly without treating a truncated tail as an error.
-fn read_chunk<T: for<'de> Deserialize<'de>>(
-    bytes: &[u8],
-    cursor: &mut usize,
-) -> Result<Option<T>, DecodeError> {
-    if *cursor >= bytes.len() {
-        return Ok(None);
-    }
-    if bytes.len() - *cursor < 4 {
-        // Partial length prefix — truncated tail.
-        return Ok(None);
-    }
-    let len = u32::from_le_bytes([
-        bytes[*cursor],
-        bytes[*cursor + 1],
-        bytes[*cursor + 2],
-        bytes[*cursor + 3],
-    ]) as usize;
-    *cursor += 4;
-    let available = bytes.len() - *cursor;
-    if available < len {
-        // Partial payload — truncated tail.
-        return Ok(None);
-    }
-    let payload = &bytes[*cursor..*cursor + len];
-    *cursor += len;
-    let (value, _): (T, usize) = bincode::serde::decode_from_slice(payload, bincode_config())
-        .map_err(|e| DecodeError::Bincode(e.to_string()))?;
-    Ok(Some(value))
-}
-
-/// Errors that can occur while decoding a `.rec` binary file.
-#[derive(Debug, thiserror::Error)]
-pub enum DecodeError {
-    /// The file ended before the named section was fully read.
-    #[error("record file truncated at {0}")]
-    Truncated(&'static str),
-    /// The 16-bit version tag in the file does not equal [`FORMAT_VERSION`].
-    #[error("format version mismatch: file={file} client={client}")]
-    VersionMismatch { file: u16, client: u16 },
-    /// A length-prefixed chunk failed bincode deserialization; carries the error message.
-    #[error("bincode decode failed: {0}")]
-    Bincode(String),
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn round_trip_header() {
-        let h = EpisodeRecordHeader {
-            format_version: FORMAT_VERSION,
-            run_id: RunId("xyz".into()),
-            seed: 42,
-            env_family: EnvFamily::Classic,
-            created_at: 1700,
-            trial: None,
-            kind: EpisodeKind::Evaluation,
-        };
-        let bytes = bincode::serde::encode_to_vec(&h, bincode_config()).unwrap();
-        let (decoded, _): (EpisodeRecordHeader, usize) =
-            bincode::serde::decode_from_slice(&bytes, bincode_config()).unwrap();
-        assert_eq!(h, decoded);
-        assert_eq!(decoded.kind, EpisodeKind::Evaluation);
-    }
-
-    #[test]
-    fn decode_rejects_version_mismatch() {
-        let mut bytes = vec![0u8; 16];
-        bytes[0..2].copy_from_slice(&999u16.to_le_bytes());
-        let err = decode_episode_record(&bytes).unwrap_err();
-        match err {
-            DecodeError::VersionMismatch { file: 999, client } if client == FORMAT_VERSION => {}
-            other => panic!("expected version mismatch, got {other:?}"),
-        }
-    }
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub enum ObjectiveSense {
+    /// Lower is better (cost, loss, error, distance-to-target).
+    Minimize,
+    /// Higher is better (reward, fitness, accuracy, score).
+    Maximize,
 }
