+++ crates/rlevo-benchmarks/src/record/mod.rs
@@ -68,10 +68,9 @@ pub use tracing_layer::RecordingLayer;
 
 pub use manifest::RunManifest;
 pub use schema::{
-    Box2dPayload, CheckpointFormat, CheckpointKind, CheckpointRef, Classic2DPayload, EnvFamily,
-    EpisodeKind, EpisodeRecord, EpisodeRecordHeader, FORMAT_VERSION, FamilyPayload, FrameRecord,
-    GridPayload, Hyperparameters, Landscape2DPayload, Locomotion2DPayload, MIN_SUPPORTED_VERSION,
-    MetricSample, PopulationSample, RecordedEnvFamily, RunId, TabularPayload, TrialRef,
+    CheckpointFormat, CheckpointKind, CheckpointRef, EnvFamily, EpisodeKind, EpisodeRecord,
+    EpisodeRecordHeader, FORMAT_VERSION, FamilyPayload, FrameRecord, Hyperparameters,
+    MIN_SUPPORTED_VERSION, MetricSample, PopulationSample, RecordedEnvFamily, RunId, TrialRef,
     bincode_config, default_frame_stride,
 };
 pub use writer::{
