+++ crates/rlevo-scene/src/lib.rs
@@ -93,17 +93,25 @@
 extern crate alloc;
 
 pub mod ascii;
+pub mod codec;
 pub mod palette;
 pub mod payload;
+pub mod schema;
 pub mod styled;
 
 pub use ascii::{AsciiRenderable, AsciiRenderer};
+pub use codec::{DecodeError, RecordChunk, decode_episode_record};
 pub use payload::{
     BodyKind, Box2dPayloadSource, Box2dSnapshot, CardTable, Classic2DBody, Classic2DPayloadSource,
     Classic2DRole, Classic2DSnapshot, GridAgentMarker, GridColor, GridDir, GridDoorState,
     GridPayloadSource, GridSnapshot, GridTile, Landscape2DPayloadSource, Landscape2DSnapshot,
     Locomotion2DPayloadSource, Locomotion2DSnapshot, Point2, RigidBody2D, TabularCell, TabularGrid,
     TabularLayout, TabularMarker, TabularMarkerKind, TabularPayloadSource, TabularSnapshot,
+};
+pub use schema::{
+    CheckpointFormat, CheckpointKind, CheckpointRef, EnvFamily, EpisodeKind, EpisodeRecord,
+    EpisodeRecordHeader, FORMAT_VERSION, FamilyPayload, FrameRecord, Hyperparameters,
+    MIN_SUPPORTED_VERSION, MetricSample, PopulationSample, RunId, TrialRef, bincode_config,
 };
 pub use styled::{Color, Modifier, SpanStyle, StyledFrame, StyledLine, StyledSpan};
